move dashboard survey methods to surveyService, add warning for deleting survey
